made the 2 pages components for easier clean up
