+++ src/router.js
@@ -13,8 +13,8 @@ import ManageExperiencesTasks from "./views/ManageExperiencesTasks.vue";
 import UserInfoDialog from "./components/UserInfoDialog.vue";
 import AdminShop from "./views/AdminShop.vue";
 import AddReward from "./components/AddReward.vue";
-import AdminViewRewards from "./views/AdminViewRewards.vue";
-import PurchaseRewards from "./views/PurchaseRewards.vue";
+import AdminViewRewards from "./components/AdminViewRewards.vue";
+import PurchaseRewards from "./components/PurchaseRewards.vue";
 import ManageUsers from "./views/ManageUsers.vue"
 import approval from "./views/approval.vue";
 import AdminBadge from "./views/AdminBadge.vue";
